fix: quality sweep addressing 13 issues (#52,#58,#59,#66,#69,#71,#72,#73,#74,#78,#88,#89,#90)
Split exercise_objectives.py (583 lines) into _types.py + per-category
objective modules (#90). Refactor body_model.py (556->165 lines) by
extracting axial_skeleton.py and foot_contact.py, delegating to
limb_builders.py (#88,#89). Enforce 300-line module budget (#78).

Fix LoD: rename _attach_barbell_to_hands and _set_floor_pull_initial_pose
to public names, eliminating cross-module private imports (#74,#72,#73).

Remove noqa B027 on hook methods -- annotated as intentional no-op hooks,
B027 false positive documented in-line (#66).

Add ensure_coordinates_within_bounds() DbC postcondition to validate
coordinate defaults stay within declared XML ranges (#52).

Document pelvis cylinder-derived radius usage for rectangular prism
inertia as intentional design choice (#58,#59).

Add 5 Rust unit tests (inverse_dynamics, COM, interpolation) (#69).
Add 20 integration tests for gait and sit-to-stand models (#71).

DRY: Remove duplicate contact helper functions from xml_helpers.py,
canonical source is contact_helpers.py.
