+++ src/opensim_models/shared/body/__init__.py
@@ -4,10 +4,8 @@ Provides a simplified but anatomically grounded full-body model with
 major body segments and joints suitable for barbell exercise simulation.
 """
 
-from opensim_models.shared.body.body_model import (
-    BodyModelSpec,
-    add_foot_contact_spheres,
-    create_full_body,
-)
+from opensim_models.shared.body._segment_data import BodyModelSpec
+from opensim_models.shared.body.body_model import create_full_body
+from opensim_models.shared.body.foot_contact import add_foot_contact_spheres
 
 __all__ = ["BodyModelSpec", "add_foot_contact_spheres", "create_full_body"]
